Code/Facial Wallet/requirements.txt
flask = "*"
flask-sqlalchemy = "*"
pillow = "*"
numpy = "*"
face-recognition = "*"
opencv-python = "*"
